extract-update-expert-location-move: Move update_expert_location_with_recovery to eplb.expert_location_updater (cut+paste)
Refactor chain ID: extract-update-expert-location-move
