+++ python/sglang/srt/eplb/expert_location_updater.py
@@ -18,6 +18,7 @@ from typing import Dict, List, Optional, Tuple
 import einops
 import torch
 import torch.distributed
+from torch import nn
 from torch.distributed import P2POp
 
 from sglang.srt.elastic_ep.elastic_ep import ElasticEPStateManager
@@ -640,3 +641,55 @@ def _group_by(items, keyfunc):
     for item in items:
         ans[keyfunc(item)].append(item)
     return dict(ans)
+
+
+def update_expert_location_with_recovery(
+    *,
+    expert_location_updater: ExpertLocationUpdater,
+    model: nn.Module,
+    new_expert_location_metadata: ExpertLocationMetadata,
+    update_layer_ids: List[int],
+    nnodes: int,
+    tp_rank: int,
+    expert_backup_client,
+    update_weights_from_disk_callable,
+    ep_dispatch_algorithm: str,
+    init_lplb_solvers_callable,
+):
+    p2p_missing_logical_experts = expert_location_updater.update(
+        model.routed_experts_weights_of_layer,
+        new_expert_location_metadata,
+        update_layer_ids=update_layer_ids,
+        nnodes=nnodes,
+        rank=tp_rank,
+    )
+
+    if len(p2p_missing_logical_experts) > 0:
+        # Load the missing expert weights from disk
+        if callable(getattr(model, "generate_weight_name_filter", None)):
+            # Filter and load only missing expert weights
+            weight_name_filter = model.generate_weight_name_filter(
+                p2p_missing_logical_experts
+            )
+        else:
+            # Do a full reload from disk/DRAM
+            logger.info(
+                "[Elastic EP] Model does not implement generate_weight_name_filter. "
+                "Performing full weight reload."
+            )
+            weight_name_filter = None
+
+        if expert_backup_client is not None and expert_backup_client.use_backup:
+            # Load the missing weights from the DRAM backup
+            expert_backup_client.update_weights(weight_name_filter)
+        else:
+            # Load the missing weights from disk
+            update_weights_from_disk_callable(
+                get_global_server_args().model_path,
+                get_global_server_args().load_format,
+                weight_name_filter=weight_name_filter,
+            )
+
+    # Re-init LPLB solvers after expert location update
+    if ep_dispatch_algorithm == "lp":
+        init_lplb_solvers_callable()
