+++ python/sglang/srt/model_executor/model_runner.py
@@ -25,7 +25,7 @@ import threading
 import time
 from collections import defaultdict
 from dataclasses import dataclass
-from typing import Any, List, Optional, Union
+from typing import Any, Optional, Union
 
 import torch
 import torch.distributed as dist
@@ -97,7 +97,6 @@ from sglang.srt.eplb.expert_distribution import (
     set_global_expert_distribution_recorder,
 )
 from sglang.srt.eplb.expert_location import (
-    ExpertLocationMetadata,
     broadcast_global_expert_location_metadata,
     compute_initial_expert_location_metadata,
     get_global_expert_location_metadata,
@@ -1642,58 +1641,6 @@ class ModelRunner(ModelRunnerKVCacheMixin):
             )
             set_global_lplb_solver(lid, solver)
         logger.info(f"Initialized LPLB solvers for {metadata.num_layers} layers")
-
-    @staticmethod
-    def update_expert_location_with_recovery(
-        *,
-        expert_location_updater: ExpertLocationUpdater,
-        model: nn.Module,
-        new_expert_location_metadata: ExpertLocationMetadata,
-        update_layer_ids: List[int],
-        nnodes: int,
-        tp_rank: int,
-        expert_backup_client,
-        update_weights_from_disk_callable,
-        ep_dispatch_algorithm: str,
-        init_lplb_solvers_callable,
-    ):
-        p2p_missing_logical_experts = expert_location_updater.update(
-            model.routed_experts_weights_of_layer,
-            new_expert_location_metadata,
-            update_layer_ids=update_layer_ids,
-            nnodes=nnodes,
-            rank=tp_rank,
-        )
-
-        if len(p2p_missing_logical_experts) > 0:
-            # Load the missing expert weights from disk
-            if callable(getattr(model, "generate_weight_name_filter", None)):
-                # Filter and load only missing expert weights
-                weight_name_filter = model.generate_weight_name_filter(
-                    p2p_missing_logical_experts
-                )
-            else:
-                # Do a full reload from disk/DRAM
-                logger.info(
-                    "[Elastic EP] Model does not implement generate_weight_name_filter. "
-                    "Performing full weight reload."
-                )
-                weight_name_filter = None
-
-            if expert_backup_client is not None and expert_backup_client.use_backup:
-                # Load the missing weights from the DRAM backup
-                expert_backup_client.update_weights(weight_name_filter)
-            else:
-                # Load the missing weights from disk
-                update_weights_from_disk_callable(
-                    get_global_server_args().model_path,
-                    get_global_server_args().load_format,
-                    weight_name_filter=weight_name_filter,
-                )
-
-        # Re-init LPLB solvers after expert location update
-        if ep_dispatch_algorithm == "lp":
-            init_lplb_solvers_callable()
 
     def maybe_recover_ep_ranks(self):
         # TODO(perf): `active_ranks.all()` on a CUDA tensor triggers host-device
